[Tizen] Make TextureCoordinate relative shader codes use highp
Let we make texture coordinates and relative shader codes use highp instead of mediump.
It will reduce visual defect for high-resolution UI system

+++ dali-scene3d/internal/graphics/shaders/default-physically-based-shader.vert
@@ -44,9 +44,9 @@ uniform int uBlendShapeGeometryHeight;
 #endif
 #endif
 
-OUTPUT mediump vec2 vUV;
+OUTPUT highp vec2 vUV;
 OUTPUT highp mat3 vTBN;
-OUTPUT lowp vec4 vColor;
+OUTPUT highp vec4 vColor;
 OUTPUT highp vec3 vPositionToCamera;
 
 uniform highp mat4 uViewMatrix;
