+++ dali-toolkit/internal/graphics/shaders/bubble-emitter.frag
@@ -1,6 +1,6 @@
 precision highp float;
 uniform vec3 uHSVDelta;
-varying mediump vec2 vTexCoord;
+varying highp vec2 vTexCoord;
 uniform sampler2D sTexture;
 
 float rand(vec2 co)
