+++ dali-toolkit/internal/graphics/shaders/emboss-filter-shader.frag
@@ -1,5 +1,5 @@
 precision highp float;
-varying mediump vec2 vTexCoord;
+varying highp vec2 vTexCoord;
 uniform sampler2D sTexture;
 uniform vec2 uTexScale;
 uniform vec3 uCoefficient;
