+++ dali-toolkit/internal/graphics/shaders/primitive-visual-shader.frag
@@ -1,7 +1,7 @@
 //Very simple fragment shader that merely applies the vertex shading to the color at each fragment.
 
-precision mediump float;
-varying mediump vec3 vIllumination;
+precision highp float;
+varying highp vec3 vIllumination;
 uniform lowp vec4 uColor;
 
 void main()
